expand config to avoid edits to source files (js and php) relocate config to common location
